+++ changelog.md
@@ -4,6 +4,8 @@ Codion Change Log
 ## 0.17.19-SNAPSHOT
 ### is.codion.common.db
 - AbstractConnectionPoolWrapper, prevent destroyed connection counter being called twice for each closed connection, see https://github.com/brettwooldridge/HikariCP/issues/2152.
+### is.codion.framework.domain
+- Condition.customCondition() renamed custom().
 
 ## 0.17.18
 - Text.padString() deprecated for removal along with Text.ALIGNMENT, Text.leftPad() and rightPad() added.